// src/base/common/enums/app.enum.ts
export enum DeviceMinSize {
    DESKTOP = 1024,
    TABLET = 768,
    MOBILE = 320,
}

export enum EHttpMethod {
    GET = 'GET',
    POST = 'POST',
    PUT = 'PUT',
    PATCH = 'PATCH',
    DELETE = 'DELETE',
}

export enum EToken {
    ACCESS_TOKEN = 'ACCESS_TOKEN',
    REFRESH_TOKEN = 'REFRESH_TOKEN'
}

export enum ELocale {
    VIETNAMESE = 'vi',
    ENGLISH = 'en',
}